Move action validation to its own file.
